+++ install.yml
@@ -1,0 +1,11 @@
+$schema: http://json-schema.org/draft-07/schema
+$id: https://adobe.io/schemas/app-builder-templates/1
+ 
+categories:
+  - code
+ 
+extension:
+  serviceCode: dx/asset-compute/worker/1
+
+services:
+- code: AssetComputeSDK
