+++ test/index.test.js
@@ -17,7 +17,7 @@ const path = require('path')
 const assert = require('yeoman-assert')
 const yaml = require('js-yaml')
 
-const { extension: { assetComputeWorker: AssetComputeWorker } } = require('../../../../index')
+const AssetComputeWorker = require('../index')
 const Generator = require('yeoman-generator')
 
 const composeWith = jest.spyOn(Generator.prototype, 'composeWith')
